# src/DocSwagger.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/OGLANEUX/portfolio/1.0.0
info:
  version: "1.0.0"
  title: portfolio Océane GLANEUX
  description: The API for the portfolio
  contact:
    name: Océane GLANEUX
    email: o.glaneux@gmail.com
paths:
...
  /Skill:
    get:
      tags:
        - Skill
      summary: Retrieve all Skill
      description: Retrieve a list of all Skill.
      operationId: getAllSkills
      responses:
        '200':
          description: OK. The list of Skill has been retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Skill'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Skill
      summary: Create a Skill
      description: Create a new Skill.
      operationId: CreateSkill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Skill'
      responses:
        '201':
          description: SkillCategory Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillCategory'
            application/xml:
              schema:
                $ref: '#/components/schemas/SkillCategory'
            application/x-www-form-urlencoded:
              schema:
                $ref: '#/components/schemas/SkillCategory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /Skill/{id}:
    get:
      tags:
        - Skill
      summary: Retrieve a Skill
      description: Retrieve details about a specific Skill by its ID.
      operationId: getASkill
      parameters:
        - name: id
          in: path
          example: 12345    
          required: true
          description: ID of the Skill to retrieve
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK. The Skill has been retrieved successfully.
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/Skill'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
          
    put:
      tags:
        - Skill
      summary: Update a Skill
      description: Update an existing Skill.
      operationId: ModifySkill
      parameters:
        - name: id
          in: path
          example: 12345    
          required: true
          description: ID of the Skill to retrieve
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Skill'
          application/xml:
            schema:
              $ref: '#/components/schemas/Skill'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Skill'
      responses:
        '201':
          description: Skill Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
            application/xml:
              schema:
                $ref: '#/components/schemas/Skill'
            application/x-www-form-urlencoded:
              schema:
                $ref: '#/components/schemas/Skill'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'       

    delete:
      tags:
        - Skill
      summary: Delete a Skill
      description: Delete an existing Skill.
      operationId: DeleteSkill
      parameters:
        - name: id
          in: path
          example: 12345    
          required: true
          description: ID of the Skill to retrieve
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: Skill deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  responses:
    NotFound:
      description: The specified resource was not found
...
    BadRequest:
      description: Bad request. The request could not be understood by the server due to malformed syntax, missing required parameters, or invalid data.
    Forbidden:
      description: The server understood the request, but refuses to authorize it. This may be due to insufficient permissions.
    InternalServerError:
      description: An unexpected error occurred on the server.
  schemas:
...
    Skill:
      type: object
      required:
        - id
        - name
        - rating
        - created_at
        - updated_at      
      properties:
        id:
          type: integer
          format: int64 
          example: 1526
          description: "The unique identifier of the skill."
        name:
          type: string
          example: "PHP"
          description: "The name the skill."
        logo:
          type: string
          example: "php_logo.png"
          description: "The logo of the skill."
        rating:
          type: number
          nullable: true
          minimum: 0
          maximum: 100
          example: 75
          description: "The proficiency rating of the skill, ranging from 0 to 100."
        skillCategory:
          type: number
          format: int64
          example: 12345
          description: "ID of the skill category. Referenced from SkillCategory."
        created_at:
          type: string
          format: date-time
          example: "2022-01-15T13:45:30Z"
          description: "The date and time when the skill was created."
        updated_at:
          type: string
          format: date-time
          example: "2022-03-20T10:30:15Z"
          description: "The date and time when the skill was last updated."
    SkillCategory:
      type: object
      required: 
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
          example: 1526
          description: "The unique identifier of the skill category."
        name:
          type: string
          example: "Front"
          description: "The name of the skill category."
        created_at:
          type: string
          format: date-time
          example: "2022-01-15T13:45:30Z"
          description: "The date and time when the skill category was created."
        updated_at:
          type: string
          format: date-time
          example: "2022-03-20T10:30:15Z"
          description: "The date and time when the skill category was last updated."
...